feat: migrate media service to MediaLibraryService for android auto
Android Auto needs a MediaLibraryService with a browsable tree, not the
MediaSessionService used for the lockscreen. Change PluginMediaService's
base class and serve the audiobook browse tree through a new
PluginLibrarySessionCallback: it answers onGetLibraryRoot/onGetChildren/
onGetItem from AudiobookBrowseTree, and translates a head unit picking a
chapter into a seek on the already-loaded player timeline.

The publication is threaded from AudiobookNavigator through the facade so
the callback can build the tree. The manifest gains the automotive
descriptor and the MediaLibraryService intent action.

Splitting the callback and the forwarding player out of the service file
keeps each class to one responsibility and the service under the size
limit.
