+++ flureadium/android/src/main/kotlin/dev/mulev/flureadium/PluginMediaService.kt
@@ -21,32 +21,23 @@ import android.content.Intent
 import android.content.ServiceConnection
 import android.content.pm.ServiceInfo
 import android.os.Build
-import android.os.Bundle
 import android.os.IBinder
 import android.util.Log
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
 import androidx.core.app.ServiceCompat
 import androidx.core.content.ContextCompat
-import androidx.media3.common.ForwardingSimpleBasePlayer
-import androidx.media3.common.Player
-import androidx.media3.common.Tracks
 import androidx.media3.common.util.UnstableApi
-import androidx.media3.session.CommandButton
 import androidx.media3.session.DefaultMediaNotificationProvider
+import androidx.media3.session.MediaLibraryService
+import androidx.media3.session.MediaLibraryService.MediaLibrarySession
 import androidx.media3.session.MediaSession
-import androidx.media3.session.MediaSessionService
-import androidx.media3.session.SessionCommand
-import androidx.media3.session.SessionResult
-import com.google.common.util.concurrent.Futures
-import com.google.common.util.concurrent.ListenableFuture
 import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.SupervisorJob
-import kotlinx.coroutines.async
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -57,104 +48,32 @@ import kotlinx.coroutines.flow.sample
 import org.readium.navigator.media.common.Media3Adapter
 import org.readium.navigator.media.common.MediaNavigator
 import org.readium.r2.shared.ExperimentalReadiumApi
+import org.readium.r2.shared.publication.Publication
 
 @OptIn(ExperimentalReadiumApi::class)
 typealias AnyMediaNavigator = MediaNavigator<*, *, *>
 
 private const val TAG = "Flutter_Readium.MediaService"
 
-private const val CUSTOM_COMMAND_REWIND_ACTION_ID = "REWIND_CUSTOM"
-private const val CUSTOM_COMMAND_FORWARD_ACTION_ID = "FORWARD_CUSTOM"
 private const val STARTUP_NOTIFICATION_CHANNEL_NAME = "Media playback"
 private const val STARTUP_NOTIFICATION_TITLE = "Flureadium playback"
 private const val STARTUP_NOTIFICATION_TEXT = "Preparing playback"
 
-@UnstableApi
-enum class NotificationPlayerCustomCommandButton(
-    val customAction: String,
-    val commandButton: CommandButton,
-) {
-    REWIND(
-        customAction = CUSTOM_COMMAND_REWIND_ACTION_ID,
-        commandButton = CommandButton.Builder(CommandButton.ICON_SKIP_BACK)
-            .setDisplayName("Rewind")
-            .setSlots(CommandButton.SLOT_BACK)
-            .setSessionCommand(SessionCommand(CUSTOM_COMMAND_REWIND_ACTION_ID, Bundle()))
-            .setCustomIconResId(androidx.media3.session.R.drawable.media3_icon_skip_back)
-            .build(),
-    ),
-    FORWARD(
-        customAction = CUSTOM_COMMAND_FORWARD_ACTION_ID,
-        commandButton = CommandButton.Builder(CommandButton.ICON_SKIP_FORWARD)
-            .setDisplayName("Forward")
-            .setSlots(CommandButton.SLOT_FORWARD)
-            .setSessionCommand(SessionCommand(CUSTOM_COMMAND_FORWARD_ACTION_ID, Bundle()))
-            .setCustomIconResId(androidx.media3.session.R.drawable.media3_icon_skip_forward)
-            .build(),
-    );
-}
-
 @ExperimentalCoroutinesApi
 @OptIn(ExperimentalReadiumApi::class)
 @androidx.annotation.OptIn(UnstableApi::class)
-class PluginMediaService : MediaSessionService(), MediaSession.Callback {
+class PluginMediaService : MediaLibraryService() {
 
     class Session(
         val navigator: AnyMediaNavigator,
-        val mediaSession: MediaSession,
+        val mediaSession: MediaLibrarySession,
+        val publication: Publication?,
     ) {
         val coroutineScope = CoroutineScope(SupervisorJob() + Dispatchers.Main)
     }
 
-    private val notificationPlayerCustomCommandButtons =
-        NotificationPlayerCustomCommandButton.entries.map { command -> command.commandButton }
-
-    override fun onConnect(
-        session: MediaSession,
-        controller: MediaSession.ControllerInfo
-    ): MediaSession.ConnectionResult {
-        val connectionResult = super.onConnect(session, controller)
-        val availableSessionCommands = connectionResult.availableSessionCommands.buildUpon()
-
-        /* Registering custom player command buttons for player notification. */
-        notificationPlayerCustomCommandButtons.forEach { commandButton ->
-            commandButton.sessionCommand?.let(availableSessionCommands::add)
-        }
-
-        return MediaSession.ConnectionResult.accept(
-            availableSessionCommands.build(),
-            connectionResult.availablePlayerCommands,
-        )
-    }
-
-    override fun onPostConnect(session: MediaSession, controller: MediaSession.ControllerInfo) {
-        super.onPostConnect(session, controller)
-        if (notificationPlayerCustomCommandButtons.isNotEmpty()) {
-            /* Setting custom player command buttons to mediaLibrarySession for player notification. */
-            /* Set media-button preferences, so that skip buttons are replaces with seek */
-            session.setCustomLayout(notificationPlayerCustomCommandButtons)
-            session.setMediaButtonPreferences(notificationPlayerCustomCommandButtons)
-        }
-    }
-
-    override fun onCustomCommand(
-        session: MediaSession,
-        controller: MediaSession.ControllerInfo,
-        customCommand: SessionCommand,
-        args: Bundle
-    ): ListenableFuture<SessionResult> {
-        /* Handle custom command buttons from player notification. */
-        if (customCommand.customAction == NotificationPlayerCustomCommandButton.REWIND.customAction) {
-            CoroutineScope(Dispatchers.Main).async {
-                ReadiumReader.previous()
-            }
-        }
-        if (customCommand.customAction == NotificationPlayerCustomCommandButton.FORWARD.customAction) {
-            CoroutineScope(Dispatchers.Main).async {
-                ReadiumReader.next()
-            }
-        }
-        return Futures.immediateFuture(SessionResult(SessionResult.RESULT_SUCCESS))
+    private val libraryCallback by lazy {
+        PluginLibrarySessionCallback { binder.session.value?.publication }
     }
 
     /**
@@ -181,6 +100,7 @@ class PluginMediaService : MediaSessionService(), MediaSession.Callback {
         @OptIn(FlowPreview::class)
         fun <N> openSession(
             navigator: N,
+            publication: Publication? = null,
         ) where N : AnyMediaNavigator, N : Media3Adapter {
             Log.d(TAG, "openSession")
 
@@ -189,17 +109,21 @@ class PluginMediaService : MediaSessionService(), MediaSession.Callback {
             // Create our SimpleBasePlayer override to override some media-button mapping.
             val pluginForwardingPlayer = PluginSimpleBasePlayer(player, ReadiumReader.audioPreferences)
 
-            val mediaSession = MediaSession.Builder(applicationContext, pluginForwardingPlayer)
+            val mediaSession = MediaLibrarySession.Builder(
+                this@PluginMediaService,
+                pluginForwardingPlayer,
+                libraryCallback
+            )
                 .setSessionActivity(activityIntent)
-                .setCallback(this@PluginMediaService)
-                .setCustomLayout(notificationPlayerCustomCommandButtons)
+                .setCustomLayout(libraryCallback.commandButtons)
                 .build()
 
             addSession(mediaSession)
 
             val session = Session(
                 navigator,
-                mediaSession
+                mediaSession,
+                publication
             )
 
             sessionMutable.value = session
@@ -268,23 +192,6 @@ class PluginMediaService : MediaSessionService(), MediaSession.Callback {
         // App and service can be started again from a stale notification using
         // PendingIntent.getForegroundService, so we need to call startForeground and then stop
         // the service.
-        /* val readerRepository = (application as org.readium.r2.testapp.Application).readerRepository
-        if (readerRepository.isEmpty()) {
-            val notification =
-                NotificationCompat.Builder(
-                    this,
-                    DefaultMediaNotificationProvider.DEFAULT_CHANNEL_ID
-                )
-                    .setContentTitle("Media service")
-                    .setContentText("Media service will stop immediately.")
-                    .build()
-
-            // Unfortunately, stopSelf does not remove the need for calling startForeground
-            // to prevent crashing.
-            startForeground(DefaultMediaNotificationProvider.DEFAULT_NOTIFICATION_ID, notification)
-            ServiceCompat.stopForeground(this, ServiceCompat.STOP_FOREGROUND_REMOVE)
-            stopSelf(startId)
-        } */
 
         // Prevents the service from being automatically restarted after being killed;
         return START_NOT_STICKY
@@ -326,7 +233,7 @@ class PluginMediaService : MediaSessionService(), MediaSession.Callback {
         notificationManager.createNotificationChannel(channel)
     }
 
-    override fun onGetSession(controllerInfo: MediaSession.ControllerInfo): MediaSession? {
+    override fun onGetSession(controllerInfo: MediaSession.ControllerInfo): MediaLibrarySession? {
         return binder.session.value?.mediaSession
     }
 
@@ -399,100 +306,5 @@ class PluginMediaService : MediaSessionService(), MediaSession.Callback {
             Intent(SERVICE_INTERFACE)
                 // MediaSessionService.onBind requires the intent to have a non-null action
                 .apply { setClass(application, PluginMediaService::class.java) }
-    }
-}
-
-@UnstableApi
-class PluginSimpleBasePlayer(player: Player, val preferences: FlutterAudioPreferences) : ForwardingSimpleBasePlayer(player) {
-
-    override fun handleSeek(
-        mediaItemIndex: Int,
-        positionMs: Long,
-        seekCommand: Int
-    ): ListenableFuture<*> {
-        // NOTE: Maps seek to next/previous track, to seek forward/backward in current track.
-        if (seekCommand == COMMAND_SEEK_TO_NEXT) {
-            return super.handleSeek(mediaItemIndex, positionMs, COMMAND_SEEK_FORWARD)
-        } else if (seekCommand == COMMAND_SEEK_TO_PREVIOUS) {
-            return super.handleSeek(mediaItemIndex, positionMs, COMMAND_SEEK_BACK)
-        }
-        return super.handleSeek(mediaItemIndex, positionMs, seekCommand)
-    }
-
-    // FIX: Hacky way to fix missing COMMAND_GET_TIMELINE from TtsSessionAdapter
-    override fun getState(): State {
-        // This is a copy & override of the super implementation, due to assert on empty playlist,
-        // which Readium TTSPlayer sometimes provides during active states.
-        // See https://github.com/readium/kotlin-toolkit/pull/716
-
-        // Ordered alphabetically by State.Builder setters.
-        val state = State.Builder()
-//      val positionSuppliers = livePositionSuppliers
-        if (player.isCommandAvailable(COMMAND_GET_AUDIO_ATTRIBUTES)) {
-            state.setAudioAttributes(player.audioAttributes)
-        }
-
-        if (preferences.allowExternalSeeking) {
-            state.setAvailableCommands(player.availableCommands)
-        } else {
-            val commandsWithoutSeeking = player.availableCommands
-                .buildUpon()
-                .remove(COMMAND_SEEK_IN_CURRENT_MEDIA_ITEM)
-                .build()
-            state.setAvailableCommands(commandsWithoutSeeking)
-        }
-
-        if (player.isCommandAvailable(COMMAND_GET_CURRENT_MEDIA_ITEM)) {
-            state.setContentPositionMs { player.contentPosition }
-            state.setContentBufferedPositionMs { player.contentBufferedPosition }
-//          state.setContentBufferedPositionMs(positionSuppliers.contentBufferedPositionSupplier)
-//          state.setContentPositionMs(positionSuppliers.contentPositionSupplier)
-        }
-        if (player.isCommandAvailable(COMMAND_GET_TEXT)) {
-            state.setCurrentCues(player.currentCues)
-        }
-        //if (player.isCommandAvailable(COMMAND_GET_TIMELINE)) {
-        state.setCurrentMediaItemIndex(player.currentMediaItemIndex)
-        //}
-        state.setDeviceInfo(player.getDeviceInfo())
-        if (player.isCommandAvailable(COMMAND_GET_DEVICE_VOLUME)) {
-            state.setDeviceVolume(player.deviceVolume)
-            state.setIsDeviceMuted(player.isDeviceMuted)
-        }
-        state.setIsLoading(player.isLoading)
-        state.setMaxSeekToPreviousPositionMs(player.maxSeekToPreviousPosition)
-        state.setPlaybackParameters(player.playbackParameters)
-        state.setPlaybackState(player.playbackState)
-        state.setPlaybackSuppressionReason(player.playbackSuppressionReason)
-        state.setPlayerError(player.playerError)
-        //if (player.isCommandAvailable(COMMAND_GET_TIMELINE)) {
-        val tracks =
-            if (player.isCommandAvailable(COMMAND_GET_TRACKS))
-                player.currentTracks
-            else
-                Tracks.EMPTY
-        val mediaMetadata =
-            if (player.isCommandAvailable(COMMAND_GET_METADATA)) player.mediaMetadata else null
-        state.setPlaylist(player.currentTimeline, tracks, mediaMetadata)
-        //}
-        if (player.isCommandAvailable(COMMAND_GET_METADATA)) {
-            state.setPlaylistMetadata(player.playlistMetadata)
-        }
-        state.setPlayWhenReady(player.playWhenReady, PLAY_WHEN_READY_CHANGE_REASON_END_OF_MEDIA_ITEM)
-        state.setRepeatMode(player.repeatMode)
-        state.setSeekBackIncrementMs(player.seekBackIncrement)
-        state.setSeekForwardIncrementMs(player.seekForwardIncrement)
-        state.setShuffleModeEnabled(player.shuffleModeEnabled)
-        state.setSurfaceSize(player.surfaceSize)
-        //state.setTimedMetadata(lastTimedMetadata)
-        if (player.isCommandAvailable(COMMAND_GET_CURRENT_MEDIA_ITEM)) {
-            state.setTotalBufferedDurationMs { player.totalBufferedDuration }
-        }
-        state.setTrackSelectionParameters(player.trackSelectionParameters)
-        state.setVideoSize(player.videoSize)
-        if (player.isCommandAvailable(COMMAND_GET_VOLUME)) {
-            state.setVolume(player.volume)
-        }
-        return state.build()
     }
 }
